Move schema to one file
