+++ api/src/index.js
@@ -1,11 +1,15 @@
-import schema from './schema';
 import {ApolloServer} from 'apollo-server';
+import {resolvers, typeDefs} from './schema';
 import {sequelize} from './db';
 
-const server = new ApolloServer({schema});
+const server = new ApolloServer({
+  typeDefs,
+  resolvers
+});
 
 sequelize
   .sync()
-  .then(() =>
-    server.listen().then(({url}) => console.log(`🚀 Server ready at ${url}`))
-  );
+  .then(() => server.listen({port: process.env.PORT}))
+  .then(({url}) => {
+    console.log(`🚀 Server ready at ${url}`);
+  });
