--- api/src/schema/replacement.js
@@ -1,21 +1,0 @@
-import {gql} from 'apollo-server';
-
-export const typeDef = gql`
-  extend type Game {
-    replacements: [Replacement]
-  }
-
-  type Replacement {
-    id: ID
-    inId: String
-    outId: String
-  }
-`;
-
-export const resolvers = {
-  Game: {
-    replacements(parent) {
-      return parent.getReplacements();
-    }
-  }
-};
